Coverity travis script

+++ tools/travis/run-script.sh
@@ -15,6 +15,15 @@ check_status()
   fi
 }
 
+configure()
+{
+  ./autogen.sh
+  mkdir bin && cd bin
+  ../configure --enable-debug --enable-silent-rules --disable-warnings $1
+  make
+  cd ..
+}
+
 if [ "$TEST"x = "coverage"x ]; then
   CONFIGURE_FLAGS="--enable-gcov"
 fi
@@ -31,32 +40,34 @@ elif [ "$COMPILER"x = "clang"x ]; then
   export CXX="clang++"
 fi
 
-./autogen.sh
-mkdir bin && cd bin
 if [ -n "$COMPILER" ]; then
-  ../configure --enable-debug --enable-silent-rules --disable-warnings $CONFIGURE_FLAGS
-  make
+  configure "$CONFIGURE_FLAGS"
+fi
+
+if [ ${COVERITY_SCAN_BRANCH} == 1 ]; then
+  exit 0;
 fi
 
 if [ "$TEST"x = "install"x ]; then
+  cd bin
   sudo make install
   denemo -n -a "(d-Quit)"
 fi
 
 if [ "$TEST"x = "integration"x ]; then
+  cd bin
   make distcheck
   rm -rf denemo-*.tar.gz
 fi
 
 if [ "$TEST"x = "coverage"x ]; then
+  cd bin
   make check
 fi
 
 if [ "$TEST"x = "cppcheck"x ]; then
-  cd ..
   cppcheck --force --enable=unusedFunction --template '{file}:{line} — {severity} — {id} — {message}' -q -I include src tools
 fi
 
 #Should be the last command ran, it checks that the repository is clean after tests are ran.
 check_status
-
